requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
motor==3.7.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.1
email-validator==2.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
requests==2.31.0 